ProcessUtilities.java: change the signature so we return the command's status code, and take two Lists, one for lines written to standard output, the other (as before) for lines written to standard error.
RevisionControlSystem.java: change the parsing methods' signatures to
take Lists instead of String[]. We'll be back when Java 1.5 is properly
with us to change this to List<String>, and remove the casting everywhere
else.

BitKeeper.java, Cvs.java: cope with the fact that we now use List instead
of String[].

RevisionWindow.java: try to use the command's status code to determine
whether or not it was successful. CVS makes life difficult, as usual. When
annotating, it feels the need to output rubbish to standard error (but does
give us a usable return code). When producing a patch, it returns a rubbish
status code (0 => no changes, 1 => changes, as far as I can tell), but at
least refrains from writing anything to standard error. BitKeeper is well
behaved in both cases.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@41 5bded901-cfd8-0310-8db1-92abb6e08623