RevisionWindow.java: refactor the UI creation to consist of lots of small methods rather than one big one.
Fix the code that switches between the annotation and patch views to not
remove all the JList's MouseListeners, because JList adds its own to
implement selection! (I think now that the clever reuse of one JList is
actually quite stupid; we'd be better off if we had two separate JList
components, perhaps switching between them with CardLayout.)

Fix an exception thrown on Mac OS because it doesn't have a system selection
like X11. Why don't I remove support for the X11 selection? Because rxvt
uses it, and -- on Linux -- I use rxvt.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@49 5bded901-cfd8-0310-8db1-92abb6e08623