Patch.java: represents a patch between revisions. Initially, this is just a dummy implementation.
RevisionWindow.java: uses a single Patch between the current and the
destination revision to do the line number translation so we arrive on
the corresponding line, rather than the line with the same line number.
Of course, with the dummy implementation of Patch we have so far, nothing
should actually have changed.

TODO: yet another thing we'll have to do. We? I.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@55 5bded901-cfd8-0310-8db1-92abb6e08623