RevisionWindow.java: fixes the problem starting RevisionTool from Edit on a CVS repository. CVS can't cope with absolute paths like BitKeeper can, so we need to start the 'cvs' commands from a directory in the appropriate repository and only use relative paths. To do this, we break incoming filenames into a repository root (which we keep as a File) and a path to the file relative to that root. All commands that used to get the raw filename now get the relative name, but are started in the repository root.
ProcessUtilities.java: backQuote gains a File parameter giving the
directory the sub-process should be started in.

RevisionControlSystem.java: isLocallyModified is currently the only
method where the work is done by the implementors of this interface
rather than by RevisionWindow, so it's the only one that needs to
be passed the repository root.

BitKeeper.java, Cvs.java: track API changes.

TODO: a job done.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@70 5bded901-cfd8-0310-8db1-92abb6e08623