src/e/scm/RevisionControlSystem.java: Log the command somewhere before running it.  This way, it'll be more difficult to lose my comments when something goes wrong.  Elliott doesn't imagine any failure case other than the one he's just fixed but that's not belt-and-braces enough.  I will not suffer my comments thrown away without also throwing my toys loudly out of the pram!
This probably isn't the best place to intercept the command execution - ProcessUtilities.backQuote is certainly called elsewhere.  I'd like to do something to preserve input more widely than just in checkintool.  For example, the find dialog throwing my input history away is still annoying me.

This probably isn't the best way to log the input.  Sometimes the invoking window has been closed and stderr points at /dev/null.  A file in /tmp would be a sensible solution but it's not clear to me whether this is the invoking script's responsibility - as the edit script might imply.

Small step.  If the nagging output acts as a motivator to do something better, then that would be excellent.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@173 5bded901-cfd8-0310-8db1-92abb6e08623