src/e/scm/Patch.java: Turns a pair of revisions of a file into a set of pairs of line numbers which diff claims correspond to one another and are identical.
The patch parsing code is ripped off RevisionWindow.java's code for handling double clicks in patch-view.  This copy of the code has the job of translating a source line number so it has to start at the top of the patch and work down until it finds the line number.  The RevisionWindow code has to handle a click within a patch so it starts at a line in the patch and works up to a hunk header.  These tasks seem quite different to me but what's common is the need to understand the patch file lines.  I've separated the code that does this from the code that uses the resulting line numbers.  My intention was that PatchLineParser and parsePatch can be split out into a separate file or files and used in the RevisionWindow code to break up the differencesDoubleClickListener.mouseClicked behemoth.

I'm not sure it was such a good idea as it would mean changing the RevisionWindow code for no extra functionality - parsing the patch top-down rather than middle-up.  I don't really know how the interface should look, so I've left it half-done, in a real YAGNI state.  On the other hand:

o  we get plus-or-minus one errors each time we implement this
o  we get confused by the back-end differences command's insistence on ordering the files the wrong way round (for translations into the past)
o  we never bother to check that we've really understood the patch properly - whereas, if we write it once, we can afford to bother with an else clause that throws an exception.

There are at least two other things which need doing:

o  handling failure of the back-end differences command
o  executing said command in the background

Translating the oldest revision 1.1 line in xray/fsb/ntfs/NtfsConsole.cpp took 20s with the GUI locked up in the current implementation.  And that's a BitKeeper repository - where the diffs are local and in the Linux filesystem cache.

src/e/scm/RevisionWindow.java: Attempting to translate the first EXTRA_DEFINES line in misc/makerules/vars-not-already-included.make from r1.104 back to its previous modification in r1.33 demonstrated that translating the line numbers through patches needs to be done step-by-step through the revision history in certain cases, as I originally feared.

I've saved a copy of the two revisions and the pathological patch between the two (but not checked them in because of intellectual property considerations).  Diff has decided that, although the line itself hasn't changed, it should be in both the red minus-prefixed lines and the blue plus-prefixed lines.  Worse still, there are black lines in between the two.  Hmm... how to explain it?  Lucidity continues to escape me!  It's like diff's saying that line 2 translates to line 2 but line 1 translates to line 3.  As diff doesn't do windiff-style reordering of changes, it's forced to say that line 1's been deleted and line 3 added, even though they're identical.

src/e/scm/LineMapper.java: We get a list of the black context lines from Patch and use these to translate line numbers across the revision or revisions covered by the patch.  The process of translation is:

(1) Translate the line directly if it was a context line.  The patch directly gives us the correspondence between the context lines so this is trivial.

(2) Determine if the line to be translated is before the patch.  If so, then the translation is the identity transformation.

(3) Determine if the line is after the patch.  If so, then the translation consists of adding the offset between the files at the end of the patch.

(4) If the line is within the patch and not a context line then it's a from-line or a to-line, a red line or a blue line, a plus line or a minus line.  The most obvious circumstance where this arises is where the line doesn't have a corresponding identical line on the other side of the patch.  However, experiments today prove that it may do, at least in the case where the patch spans more than one revision.

If we encounter such an ambiguity, we make an educated guess as to where in the file the user would like to end up and print a warning.  This may or may not be more annoying than jumping back to the start.  Hopefully the step-by-step translation in RevisionWindow.java will mean that it doesn't happen.  With non-linear BitKeeper history, though, I fear it will.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@197 5bded901-cfd8-0310-8db1-92abb6e08623