When we select a revision, we shouldn't lose the information the user has carefully given us about which line they're interested in.  We already have the code to translate line numbers between revisions.
src/e/scm/AnnotatedLineRenderer.java: 
src/e/scm/RevisionWindow.java:

getSelectedRevision / getAnnotatedRevision:  What we're really interested in when we ask for the selected revision is the revision that we're annotating - we (should) never ask for it when we're looking at a differences view.  (I say "(should)" because we could ask for a ChangeSet when in differences view - I've prevented that now.)  So "getAnnotatedRevision" is a better name.

src/e/scm/RevisionWindow.java:

translateLineNumberStepByStep:  If we try to translate a line number in an older revision to a newer one, we run into the problem of which order to iterate over the revisions.  Rather than try to solve this - which wouldn't be difficult (though it's not obvious how to do it tastefully enough that we don't have to revisit it at some point in the near future).

getAnnotatedRevision:  We need to store this separately from the selection revision so that, when the selected revision changes, we can tell what it changed from, so we can translate the line number from one to the other.

Make it selectRevision's job to select the line number as well as the revision.  That way, we avoid having two BackEndWorkers - from two calls to showAnnotationsForRevision - racing to set the line number.  The one created by the RevisionListSelectionListener would try to set the line number to 0 or (now) to the translated current line number.  The one from selectRevision's caller would try to override this.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@230 5bded901-cfd8-0310-8db1-92abb6e08623