src/e/scm/RevisionControlSystem.java: support for creating files containing the comment or the list of files to be checked in. We need these to work around OS command-line limits that have been causing me trouble on Linux at work. I had thought it was something to do with blocking by not reading stderr from the pipe, but fixing that had no effect. And I did find that interrupting a hung "bk delta" and then telling CheckInTool to commit only those files whose status hadn't changed from 'M' to '+' did exactly what I wanted. (The only bit I can't explain is why "bk delta" hung.)
src/e/scm/BitKeeper.java: for BitKeeper, run "bk delta" once for each file, and
don't give a list of files to "bk commit", because Google suggests that isn't
necessary. Google doesn't know how to get the check-in comment from a file, so
that's still limited by command-line length.

src/e/scm/Cvs.java: CVS, conversely, has a well-known option to get the
check-in comment from a file, but doesn't appear to have any mechanism to get
the list of files to be checked in from anywhere but the command line. So here
the limit isn't on the size of the comment; it's on the size of the filenames.

src/e/scm/Subversion.java: Subversion has well-known options for both, so it
should now be free of limitations. (Assuming this commit works!)



git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@251 5bded901-cfd8-0310-8db1-92abb6e08623