TODO-checkintool: Fix my repeated gripe about the throwing away of the check-in comment by...
src/e/scm/CheckInWindow.java: ... having the same bit of code that enables or disables the check-in comment area being responsible for loading the saved comment.  That way, rather than it clearing the comment after it's been loaded, we always switch from the help to the saved comment.  This also means that, if you deselect the last file, you haven't also thrown away your whole comment.  (You can get it back by selecting a file again.  I think that's a feature but I'm not attached to it.)

I could have made the tableChanged hunk smaller but I prefer this, more symmetrical version (which still carefully preserves the read-only state of the instructions).  In updateFileStatuses, we now need to save the comment as well as the list of files.  That's a good thing - I like having my precious input in stable storage.  Move the hack to cope with a null model at the time of the first call into updateFileStatuses where it also protects us from overwriting the saved comment with the contents of the not-yet initialised checkInCommentArea.

While looking for other callers of checkInCommentArea.setEnabled that might need changing, I found clearCheckInCommentArea.  That method is really saying "we're done with the user's previous input - it's safely committed - throw it away".  It wasn't explicitly throwing away the list of included files.  Perhaps, then, if the user quit during the updateFileStatuses which follows, then a restarted checkintool would have used the old list.  That's not right, although it's only a cosmetic issue.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@264 5bded901-cfd8-0310-8db1-92abb6e08623