src/e/scm/CheckInWindow.java: Previously, I just filled the check-in comment area when the table status changed from having a file selected to having none selected or vice versa.  That was sufficient to fill it with the saved comment if there were saved selected files.  It wasn't sufficient to fill it with the help if there was nothing saved.  I had a lot of chicken-and-egg trouble with this code before.  So I'm suspicious that the first thing I tried appears to have worked this evening, restarting with no saved files and with.
The BlockingWorker's finish() method, which is where the new call belongs, is (I see) run on the Swing thread.  Presumably the TableModelListener is too, although the JavaDoc doesn't make it easy to discover that.  Who cares which thread your callbacks get invoked on?  Duh, everyone.  It might well be that all Swing callbacks are executed on the Swing thread but where's the link to the article explaining that?


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@292 5bded901-cfd8-0310-8db1-92abb6e08623