src/e/scm/CheckInWindow.java: A more elegant search for the existing "filename: " label.  Elliott suggested something like:
checkInComment.matches("(?m)^" + StringUtilities.regularExpressionFromLiteral(name))

But I ran into the sand of multi-line regular expressions and "matches" wanting the whole string to be matched.

While I'm here, fix the search so it doesn't insist on there being trailing whitespace after the colon.  Relax the check for how many newlines we need similarly - so it doesn't append two newlines if the comment ends in ":" but not ": ".  Don't append a newline if the check-in comment already ends in a blank line.  "name" previously meant "label" - name + ": ".  Dealing with the name alone is more useful if we want to be flexible about the space after the colon.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@308 5bded901-cfd8-0310-8db1-92abb6e08623