src/e/scm/CheckInWindow.java: Cope with trimmed whitespace.  Trailing whitespace, which isn't indentation, is as useless everywhere as it is in source.  Source tends to be edited later, when the trailing whitespace gets in the way.  Check-in comments don't, although they may in future be annotated.
I do agree with Elliott that it's distressing that this method (or class, if it were split) contains ":", ": " and ": ?".  Parsing text that's ambiguous isn't trivial.  That's no excuse for over-complicating it, just a sign that it's likely to be more complicated than parsing something simpler.

I agree that this would be a bad idea:

  the other alternative would be to insert ":", but any human would have to start typing with a " ", and then you'd probably want to cope with the situation where they'd removed all their visible characters on a line, but left it ending ": ".
  
And this was right too:
  
  i know a filename can contain ": " as easily as it contain ":", but i think the latter is more likely (times, for example, or Mac volume names)

But lines of human-written text can end in colon very easily too.  They could plausibly end in colon-space.  That's less likely, but as Elliott himself argues above, plausible.  We can do better than matching what the line ends with - we can see if the whole line is something that we would have added.  If the whole line is visibly indistinguishable from something we've added, then it's really difficult to see how we can do better than assume that we've added it.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@310 5bded901-cfd8-0310-8db1-92abb6e08623