src/e/scm/CheckInTool.java: src/e/scm/RevisionTool.java: Although these error messages weren't quite duplicates, they're not always accurate.  If we leave the error reporting to the level which knows what the error was...
src/e/scm/RevisionControlSystem.java: ... then we can only have one-size-fits-all messages but at least we can have a different, accurate message for each of the two failure cases (discovered so far).  We could throw something to allow the caller to customize the message but the one-size-fits-all message looks adequate to me:

martind@duezer:~$ checkintool
2005-06-01T11:00:04.270-0700 CheckInTool: Application started.
2005-06-01T11:00:04.274-0700 CheckInTool: Java 1.5.0_03.
2005-06-01T11:00:04.303-0700 CheckInTool: Linux 2.6.11-1-686-smp (i386).
/home/martind is not in a directory that is under revision control
martind@duezer:~$

There is a third caller of RevisionControlSystem.forPath: RevisionWindow's constructor.  That's only called when the RevisionControlSystem has already been found for that path.  So it will only fail if the file system has changed underneath us.  It doesn't matter if we print something on System.err in that case.  We'll dereference null shortly afterwards.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@316 5bded901-cfd8-0310-8db1-92abb6e08623