src/e/scm/RevisionView.java: src/e/scm/RevisionWindow.java: I thought I needed this refactoring (it was my motivation for the too-early previous commit), but after I had it, I realized it wasn't what I wanted. But it's probably a good change, so I'll keep it.
src/e/scm/Patch.java: remove an unused parameter.

src/e/scm/AnnotatedLineRenderer.java: this wants RevisionView, not the new RevisionWindow.

src/e/scm/Subversion.java: monkey about with the paths we get from Subversion in a way that's probably specific to how we're using it. I wonder if there's a good way to get rid of the "trunk" nonsense for those of our repositories that have it? An alternative potential solution is mentioned in the FIXME.

src/e/scm/ChangeSetWindow.java: a rough first attempt at a change set browser. This is Subversion-specific and rather unpleasant to use.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@364 5bded901-cfd8-0310-8db1-92abb6e08623