src/e/scm/ChangeSetItem.java: Make the ChangeSetItems comparable...
src/e/scm/ChangeSetWindow.java: ... in order to be able to sort the list just before inserting it into the JComboBox.  That's the latest in the call stack where I could do the sorting.  Given that I don't want to sort at the earliest point - in the back-end dependent part - then the latest seemed the best place.

I was trying to discover whether the lists produced by csettool and revisiontool were the same but I was having trouble finding a single matching file in both lists.  Unsurprising, perhaps, given that I was comparing two unsorted lists each containing ~3000 items.

martind@duezer:~/work/octopus$ bk r2c -r1.2 headers/HardwareRegister.h 
1.3
martind@duezer:~/work/octopus$ bk rset -r1.3  | wc
   2953    2953  147287
martind@duezer:~/work/octopus$ 

Having sorted the one of the lists to which I had the source, I found a couple of matching items.  That's enough to convince me that the entire list is identical (:-)).  That means that we're providing better functionality than revtool, which says that this revision of HardwareRegister.h is "Not in a CSET".  It is in a change set, it's just not the last revision of HardwareRegister.h that's in that same change set.  I've always wondered what the unhelpful "Not in a CSET" meant.  More to the point, I've often given up trying to track down a problem when revtool wouldn't show me the change set for that revision.  Another possibly deliberate but mistaken UI decision stops a user getting some information which the computer was perfectly able to provide.

Enough ranting about revtool: I should add that I'm not convinced this is the right ordering.  We might want to take into account the revisions as well as the filename.  I think that's likely, although whether we'll ever care enough to implement it, I'm less convinced of.  We might want the list to be in the same order as the files are mentioned in the check in comment.  I think that would be a bad idea for the case when you really care - when the list is large.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@381 5bded901-cfd8-0310-8db1-92abb6e08623