src/e/scm/ChangeSetWindow.java: add a "Show Patch" button.
src/e/scm/PatchView.java: separate the preparation of a patch from the showing of a patch.

src/e/scm/RevisionView.java: pass the list of revisions through to ChangeSetWindow so it can get hold of the comments without the cost of asking the back end again. Also track an API change.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@403 5bded901-cfd8-0310-8db1-92abb6e08623