src/e/scm/ChangeSetWindow.java: Create a RevisionListModel for each file in the change set, so we get the right revision numbers to pass to bk diffs.  Only enable the Show Patch button when we've got all the RevisionListModels.  This makes Subversion behavior slower linearly in the number of files in the change set for no functional benefit but fixes the BitKeeper behavior.
BackEndWorker is a copy of the nested class from RevisionWindow with the status-setting and progress indication removed.  showToolError is a direct copy from RevisionView.  Its caller - readListOfRevisions - isn't quite such a direct copy.  I've started to separate the functionality that wants to be shared between ChangeSetWindow and RevisionWindow from that which wants to be ChangeSetWindow-specific.  I haven't got very far because I'm not sure it's worth investing much effort before finding out what's going to happen to BackEndWorker.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@408 5bded901-cfd8-0310-8db1-92abb6e08623