src/e/scm/RevisionListWorker.java: This class gains a reference to the one field from RevisionView that it used to use.  It now passes along a StatusReporter - the means by which a BackEndWorker updates the UI.
src/e/scm/BackEndWorker.java: I've tied together the setting of the status and the starting of the progress animation as startTask and similarly for finishTask.  Partly I was trying to keep the interface simple and to ensure that we update both consistently.  Previously, the RevisionView code sometimes set the status to a "Task..." string without starting the animation.  That may have been intentional although I'm not convinced it was right - if we're saying "...", we're saying it may take a while.  A first class BackEndTask object, containing the handle and the message, looks like it would be an improvement.  I could have left reportException here, and passed the owner Component to BackEndWorker rather than StatusReporter.  This way meant slightly less code.

src/e/scm/ChangeSetWindow.java: 
src/e/scm/RevisionView.java: Lose BackEndWorker and RevisionListWorker.

src/e/scm/ChangeSetWindow.java: Gain a StatusReporter with no visible manifestation, to keep BackEndWorker happy.

src/e/scm/RevisionView.java: Combine the fields which report the status of the back-end jobs as StatusReporter and lose them from here.

src/e/scm/StatusReporter.java: I don't understand why we go to some lengths to ensure that the JLabel never has an empty string.  I've removed the duplicated " " from the initializer but, via setMessage, caused the " " to be restored during the StatusReporter constructor.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@420 5bded901-cfd8-0310-8db1-92abb6e08623