src/e/scm/ChangeSetWindow.java: The previous change sometimes broke showing a patch because the BackEndWorker from RevisionWindow (which is now the only BackEndWorker) assumes that only the most recently started worker has value.  All of the BackEndWorkers started by ChangeSetWindow are run in parallel and all have value.
Again, I think, this points to a BackEndTask object fighting to get out.  Normal BackEndWorkers would be passed a new BackEndTask(message).  The ChangeSetWindow's workers would all be passed the same BackEndTask.  A worker would attach itself to the task and detach itself.  The last worker to detach stops the progress animation, assuming that the BackEndTask isn't outdated.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@421 5bded901-cfd8-0310-8db1-92abb6e08623