src/e/scm/ChangeSetWindow.java: Lose the hack added in the previous revision, in favor of sharing the same BackEndTask object for all of our concurrent BackEndWorkers.
src/e/scm/RevisionView.java: This means a little more boilerplate when we want to create a new task for each worker, but nothing more than a little boilerplate.

src/e/scm/StatusReporter.java: Lose the message to the BackEndTask.

src/e/scm/BackEndWorker.java: Now we just deal with a BackEndTask, not a message and a taskHandle.  We don't need to directly interact with the StatusReporter any more.

src/e/scm/RevisionListWorker.java: Adapt to the BackEndWorker change.

src/e/scm/BackEndTask.java: As suggested in the previous check-in comment, a new class to encapsulate what, to the user, appears as a single back-end task.  Support the execution of this task in parallel by several workers.  Only report the task as finished when all the workers have finished, regardless of the order in which they finish.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@422 5bded901-cfd8-0310-8db1-92abb6e08623