src/e/scm/Subversion.java: It's possible, though not advisable for anything other than testing purposes, to allow anonymous write access to a Subversion repository. Changes made anonymously are marked as having been made by "(no author)".
Change the author's-name-matching pattern to match any character except a bar, so that revisiontool includes in its revision history list all revisions of a file, including those made anonymously. Without this change, if all revisions are anonymous the revision history list would be empty and would throw an exception instead of returning the latest revision in the repository. If some but not all revisions are anonymous it would throw an exception whenever the annotation view referred to an anonymous revision, and the corresponding entry was not present in the RevisionListModel.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@458 5bded901-cfd8-0310-8db1-92abb6e08623