src/e/scm/CheckInWindow.java: martind was seeing this kind of problem:
java.lang.IndexOutOfBoundsException: Index: 1, Size: 0
	at java.util.ArrayList.RangeCheck(ArrayList.java:547)
	at java.util.ArrayList.get(ArrayList.java:322)
	at e.scm.StatusesTableModel.getFileStatus(StatusesTableModel.java:48)
	at e.scm.CheckInWindow.updatePatchView(CheckInWindow.java:188)
	at e.scm.CheckInWindow.access$300(CheckInWindow.java:15)
	at e.scm.CheckInWindow$4.windowGainedFocus(CheckInWindow.java:108)
	at java.awt.Window.processWindowFocusEvent(Window.java:1863)
	at java.awt.Window.processEvent(Window.java:1778)
	at java.awt.Component.dispatchEventImpl(Component.java:4410)
	at java.awt.Container.dispatchEventImpl(Container.java:2116)
	at java.awt.Window.dispatchEventImpl(Window.java:2429)
	at java.awt.Component.dispatchEvent(Component.java:4240)
	at java.awt.KeyboardFocusManager.redispatchEvent(KeyboardFocusManager.java:1848)
	at java.awt.DefaultKeyboardFocusManager.typeAheadAssertions(DefaultKeyboardFocusManager.java:878)
	at java.awt.DefaultKeyboardFocusManager.dispatchEvent(DefaultKeyboardFocusManager.java:365)
	at java.awt.Component.dispatchEventImpl(Component.java:4282)
	at java.awt.Container.dispatchEventImpl(Container.java:2116)
	at java.awt.Window.dispatchEventImpl(Window.java:2429)
	at java.awt.Component.dispatchEvent(Component.java:4240)
	at java.awt.EventQueue.dispatchEvent(EventQueue.java:599)
	at e.debug.EventDispatchThreadHangMonitor.dispatchEvent(EventDispatchThreadHangMonitor.java:191)
	at java.awt.SequencedEvent.dispatch(SequencedEvent.java:98)
	at java.awt.EventQueue.dispatchEvent(EventQueue.java:597)
	at e.debug.EventDispatchThreadHangMonitor.dispatchEvent(EventDispatchThreadHangMonitor.java:191)
	at java.awt.EventDispatchThread.pumpOneEventForFilters(EventDispatchThread.java:273)
	at java.awt.EventDispatchThread.pumpEventsForFilter(EventDispatchThread.java:183)
	at java.awt.EventDispatchThread.pumpEventsForHierarchy(EventDispatchThread.java:173)
	at java.awt.EventDispatchThread.pumpEvents(EventDispatchThread.java:168)
	at java.awt.EventDispatchThread.pumpEvents(EventDispatchThread.java:160)
	at java.awt.EventDispatchThread.run(EventDispatchThread.java:121)

I reproduced this by adding a Thread.sleep in the Mercurial back-end's "commit" method to make it slow like BitKeeper or Subversion (one of which he was probably using), and could then reproduce the problem at will by having more than one modified file, committing only the last, and clicking focus away to another window as soon as I hit "Commit". The problem turned out to be the line in discardSavedState that was assigning an empty table model to the statusesTableModel field (but not actually switching the JTable's model). This was presumably so that the call to updateSavedState would remove all record of now-committed files so that the later call to updateFileStatuses wouldn't try to re-check the now-committed files. I don't think this is necessary because we should ignore them when we see that they're not in the new model.

I've made a few other changes too, where I noticed that we were doing Swing work off the EDT. We should junk the existing code and switch to SwingWorker, especially now Java 6 is just around the corner.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@468 5bded901-cfd8-0310-8db1-92abb6e08623