src/e/scm/ChangeSetWindow.java: If we request select the first file in the list while threads are still populating the list, then we now do this:
new WaitCursor(this, "Getting patch...", statusReporter)

This marks the previous, still-running task:

new BackEndTask("Getting list of revisions...", statusReporter)

... as "outdated", causing its results to be discarded.  This always leads to Show Patch remaining greyed out for me with a bk back-end and a three-file change set, although seemingly didn't cause problems for others.

Defer our selection of the first item on the list until the list is fully populated.  We already know when that is, because we only enable the Show Patch button then.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@478 5bded901-cfd8-0310-8db1-92abb6e08623