src/e/scm/CheckInWindow.java: Copy a .o file to a directory that's not ignored, start CheckInTool, select the .o file, right click, Discard Changes.  A window pops up to say:
java.lang.RuntimeException: binary file
	at e.util.StringUtilities.readFile(StringUtilities.java:29)
	at e.scm.CheckInWindow$11.work(CheckInWindow.java:329)
	at e.scm.BlockingWorker.run(BlockingWorker.java:22)
	at java.lang.Thread.run(Thread.java:595)

This is because the method we use to take a backup copy doesn't work for binary files.  That's lame but I'm not fixing that in this check-in.  In this check-in, I'm fixing the more annoying problem that, when you dismiss that monolog box, the rest of the UI remains disabled.

Now I know I'm only doing half the job here, if that, because Commit and Refresh both use also use BlockingWorker, like Discard, and both also disable the UI.  Commit looks to contain some exception handling but I submit it's clumsier than what I've done here.  I'm less sure of Refresh and whether the triplication could be collapsed.  Let's see what Elliott thinks.


git-svn-id: file:///home/martind/software.jessies.org/svnroot/scm@493 5bded901-cfd8-0310-8db1-92abb6e08623