D58880: Enforce that every checkin to a fish repository contains a defect number or backlog id
bin/checkintool: In which tis writ:

Andy Chittenden <andyc>, 2011-03-28: ...

When committing something using checkintool, if you don't set the BK_GUI variable, the prompt comes out in the window you launched checkintool from. It's not obvious that's happened as it's normally hidden by the checkintool window  and there's a lot of verbosity in that window from checkintool itself :-(. If you do notice it, then it's a stopped process and so you have to "fg" it to be able to type. If you do set the BK_GUI variable, then you get a GUI  window (as per the experience with bk citool). When you click on "reenter", you get a further java window moaning that it's failed for some reason with a backtrace. At least when you dismiss that, the checkintool window is still there for further changes.

FWIW I use bk prompt which either uses a GUI or command line prompt depending on whether BK_GUI variable is set. Ideally, I'd like to be able to detect I'm running under checkintool in the trigger and set the BK_GUI prompt automatically in the trigger but I cannot think of a way of doing that.