D63325: bk-5.2.1: revisiontool's "Show patch" fails with NullPointerException at e.scm.ChangeSetWindow.showPatch(ChangeSetWindow.java:114) aka http://db.bitkeeper.com/cgi-bin/view.cgi?.page=view&id=2011-08-23-001: rset's output format has changed between 4.6b and 5.2.1 but the help hasn't
src/e/scm/BitKeeper.java: In which I try to simulate the output of bk-4.6b's rset with bk changes.  BitMover explained that my suggestion wouldn't behave quite the same as the old code in the case where there was more than one delta of the same file in a change set.  It arguably works better than the old code for revisiontool.