[Bug] D82806: checkintool shouldn't include pending deltas unless they're explicitly selected
src/e/scm/BitKeeper.java: Only the check-in revisions of the selected files should be included in the change set, otherwise unrelated bk cp, bk mv etc operations get pushed accidentally and without warning.  This required passing the list of selected delta to bk commit on stdin...

src/e/scm/RevisionControlSystem.java: ... which meant changes to ProcessUtilities.backQuote and everything that called it.

This is the change that allows compilation with modern salma-hayek.