[RFE] D92809: multiple selection in checkintool
src/e/scm/CheckInWindow.java: In which tis writ:

I had a look in the history to see if it was deliberately constrained to single file selection.  Not as far as I saw.  I wonder what happens if we give our developers a bigger hammer with which to wreak.  It's not often that this would be beneficial and I imagine it'll encourage shift-clicking without engaging brain or looking at the intervening patches.  There's also an issue with the quality of the implementation.  When I thought it was just a matter of this:

<snipped>

Then I thought the toolkit would take care of making it work like other sensible applications.  But that just let me select multiple rows at once without altering the state of the checkbox on any row but the one I clicked last.  I've got something that might do the trick:

<snipped>

But I'm no user interface programmer.  Heck, I don't really know Java.  This might not behave like a discriminating user would expect.  Oh well, let's give it a shot.

I then asked for more on the motivation and got back a response that included this:

> I hear you Martin, that shift-click could be seen as dangerous but I've also seen some bad stuff as a result of the cli being used to mass check-in, if you're not careful in what you do (wildcards being used and the like).

Andy had a quibble:

> FWIW I first just used shift click not in the tick box and it selected them but didn't do anything else: I was a bit confused.

But I was able to satisfy him with:

The way it is in the version you're using right now, with the selection independent of the inclusion state, means that you can exclude the selection from the change set by clicking on the checkbox.  That seems pretty useful too, though I suffered the same initial confusion as Andy.

The code that removes the filenames from the check-in comment box when the file is excluded from the change set has never been completely right and the shift-clicking now lets me make a prodigious mess with it.  Oh well.

...

No-one said anything about Edit File or Show History applying to multiple selections so I did nothing.

In case of accident, if I discard a file from a repository ~martind/work/misc called makerules/badger, that file gets copied into /tmp/makerules%2Fbadger.