[Bug] D118239: Java NullPointerException thrown when attempting "revisiontool leak.rb"
src/e/scm/RevisionView.java: Perhaps because of titular problem, seen with Java 8, already fixed in salma-hayek, we were sensitized to an exception report that might have long been there.  The first click on the revision list would give:

    Associated exception:
    java.lang.IllegalArgumentException: fromRevision must be newer than toRevision or this code needs fixing
        at e.scm.RevisionView.getRevisionRange(RevisionView.java:37)
        at e.scm.RevisionView.translateLineNumberStepByStep(RevisionView.java:60)
        at e.scm.RevisionView$6.work(RevisionView.java:381)
        at e.scm.BackEndWorker.run(BackEndWorker.java:26)
        at java.lang.Thread.run(Thread.java:745)

This was seen with Java 7.  The code was trying to track the current line number step by step to the new revision, but there was no new revision.  It should have returned the current line number when the fromRevision and the toRevision are the same.