[Bug] D125484: Selecting 'Reenter comments' when prompted by BK-TRIGGER-pre-commit-police-changeset-comments.rb only re-enters comments for the changeset, not for individual files in the changeset
src/e/scm/BitKeeper.java: This file's history contains a comment, relating to "esra", containing the LOG_SEPARATOR string.  That stops revisiontool working on it.  I worried that the LOG_SEPARATOR might appear in the output that we show the user.  Apparently not: that line of hyphens is called COMMENT_SEPARATOR and is two characters longer.  Having two things that look the same but aren't is confusing!  If this code is behaving as I think it is, then the LOG_SEPARATOR will never be seen, so it doesn't have to be visually innocuous, like a line of hyphens.  We could use some non-printing character but perhaps form-feed has an appropriate purpose.  Let's combine that with an explanation, to make it less likely to appear by accident.

src/e/scm/Git.java: Of course, it's actually this (spit) duplicate of the string that I'm using when I read the history of BitKeeper.java.